Move v2 classes into Courses module
- Replace `SearchCoursesForm` with `Courses::SearchForm` for improved namespacing.
- Introduce `Courses::Query` to handle course query logic, replacing the existing `CoursesQuery`.
- Move `CoursesQuery::Logger` to `Courses::QueryLogger` for consistency with the new namespacing.
